CGS updates, allowing for switching to text mode from graphics + more changes
